fix: give zoom transition highest priority
**Description**

This makes sure that the opening/closing transitions are triggered even though an image CSS `transition` is provided as inline style.

This is problematic with frameworks that generate inline styles on their
images (e.g. Gatsby).

See full context in #110.

**Solution**

The CSS `transition` is marked as "!important" for the animation to happen
even though it's overriden by another inline `transition` style attribute.

**Related**

Closes #110